+++ src/components/goals/GoalSelect.js
@@ -1,10 +1,10 @@
-import React from 'react';
+import React, { forwardRef } from 'react';
 import { Select } from 'antd';
 import PropTypes from 'prop-types';
 import Icon from 'components/common/Icon';
 import { useGoalApi } from 'hooks/UseGoalApi';
 
-export const GoalSelect = React.forwardRef(function GoalSelect(props, ref) {
+const GoalSelect = forwardRef(function GoalSelect(props, ref) {
     const goalApi = useGoalApi();
     const value = goalApi.goals.find(goal => goal.id === props.value) ? props.value : undefined;