+++ src/components/tasks/calendar/CalendarEvent.js
@@ -4,9 +4,7 @@ import { TaskTitle } from 'components/tasks/common/TaskTitle';
 
 function CalendarEvent({ event }) {
     return (
-        <TaskTitle
-            taskId={event.task.id}
-            settings={event.settings} />
+        <TaskTitle taskId={event.task.id} />
     );
 }